+++ test/plugin/scenarios/guava-eventbus-scenario/bin/startup.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -14,6 +16,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-1.1.0
-1.2.1
-1.3.0
+home="$(
+  cd "$(dirname $0)"
+  pwd
+)"
+
+java -jar ${agent_opts} ${home}/../libs/guava-eventbus-scenario.jar &