+++ CHANGES.md
@@ -7,6 +7,7 @@ Release Notes.
 * Support Undertow thread pool metrics collecting.
 * Support Tomcat thread pool metric collect.
 * Remove plugin for ServiceComb Java Chassis 0.x
+* Add Guava EventBus plugin.
 * Support Druid Connection pool metrics collecting.
 * Support HikariCP Connection pool metrics collecting.
 * Support Dbcp2 Connection pool metrics collecting.